[Chore]: ITranscription 인터페이스 추가, IMember 인터페이스 분리
