+++ types/IGroup.ts
@@ -1,9 +1,4 @@
-export interface IMember {
-  memberId: number;
-  nickname: string;
-  image: string | null;
-  status: 'PARTICIPATE' | 'PENDING';
-}
+import { IMember } from './IMember';
 
 export interface IGroup {
   groupId: number;
