+++ types/IMember.ts
@@ -1,0 +1,6 @@
+export interface IMember {
+  memberId: number;
+  nickname: string;
+  image: string | null;
+  status: 'PARTICIPATE' | 'PENDING';
+}
